caches status also in Modulith
