+++ caching-service/src/main/java/org/zowe/apiml/caching/health/CachingHealthIndicator.java
@@ -13,20 +13,16 @@ package org.zowe.apiml.caching.health;
 import com.netflix.discovery.shared.Application;
 import jakarta.annotation.Nonnull;
 import lombok.RequiredArgsConstructor;
-import org.infinispan.spring.embedded.provider.SpringEmbeddedCacheManager;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.actuate.health.AbstractHealthIndicator;
 import org.springframework.boot.actuate.health.Health;
 import org.springframework.boot.actuate.health.Status;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
-import org.springframework.cache.CacheManager;
 import org.springframework.context.ApplicationListener;
 import org.springframework.stereotype.Component;
 import org.zowe.apiml.eurekaservice.client.ApiMediationClient;
 import org.zowe.apiml.product.constants.CoreService;
 
-import java.util.HashMap;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicReference;
 
@@ -40,51 +36,22 @@ public class CachingHealthIndicator extends AbstractHealthIndicator implements A
 
     private final AtomicReference<Boolean> serviceUp = new AtomicReference<>(false);
 
-    @Value("${caching.storage.mode:inMemory}")
-    private String storageMode;
-
     private final ApiMediationClient apiMediationClient;
-    private final CacheManager cacheManager;
-
-    private boolean updateInfinispanHealth(Health.Builder builder, SpringEmbeddedCacheManager springEmbeddedCacheManager) {
-        boolean health = true;
-
-        var nativeCacheManager = springEmbeddedCacheManager.getNativeCacheManager();
-        var status = nativeCacheManager.getStatus();
-
-        var infinispan = new HashMap<String, Object>();
-        infinispan.put("status", status);
-
-        health &= status.allowInvocations();
-        var caches = new HashMap<String, Object>();
-        for (String cacheName : cacheManager.getCacheNames()) {
-            var cacheStatus = nativeCacheManager.getCache(cacheName).getStatus();
-            caches.put(cacheName, cacheStatus);
-            health &= cacheStatus.allowInvocations();
-        }
-        infinispan.put("caches", caches);
-        builder.withDetail("infinispan", infinispan);
-
-        return health;
-    }
+    private final Optional<CachesHealthIndicator> cachesHealthIndicator;
 
     @Override
     protected void doHealthCheck(Health.Builder builder) {
-        var eurekaClient = apiMediationClient.getEurekaClient();
-        boolean gatewayUp = Optional.ofNullable(eurekaClient.getApplication(CoreService.GATEWAY.getServiceId())).map(Application::getInstances).map(i -> !i.isEmpty()).orElse(false);
-        boolean health = serviceUp.get() && gatewayUp;
+        boolean gatewayUp = Optional.ofNullable(apiMediationClient.getEurekaClient())
+            .map(eurekaClient -> eurekaClient.getApplication(CoreService.GATEWAY.getServiceId()))
+            .map(Application::getInstances)
+            .map(i -> !i.isEmpty())
+            .orElse(false);
+        builder.withDetail(CoreService.GATEWAY.getServiceId(), gatewayUp ? Status.UP : Status.DOWN);
 
-        if (
-            "infinispan".equals(storageMode) &&
-            cacheManager instanceof SpringEmbeddedCacheManager springEmbeddedCacheManager
-        ) {
-            updateInfinispanHealth(builder, springEmbeddedCacheManager);
+        cachesHealthIndicator.ifPresent(i -> i.doHealthCheck(builder));
+        if (!serviceUp.get() || !gatewayUp) {
+            builder.down();
         }
-
-        Status healthStatus = health ? Status.UP : Status.DOWN;
-        builder
-            .status(healthStatus)
-            .withDetail(CoreService.GATEWAY.getServiceId(), healthStatus.getCode());
     }
 
     @Override
